Implement interfaces for roomserver storage
